steganography_api/requirements.txt
Django===5.0.6
psycopg2-binary==2.9.9
channels===3.0.5
SQLAlchemy===2.0.23
cryptography===42.0.6
django-cors-headers===4.3.1
pytz===2024.1 
requests===2.32.2
python-magic===0.4.27
stegano===0.11.3